requirement.txt
# Core Requirements
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0

# LangChain & AI/ML
langchain==0.0.347
langchain-community==0.0.19
langchain-chroma==0.0.3
langchain-openai==0.0.5

# Vector Store & Embeddings
chromadb==0.4.15
sentence-transformers==2.2.2

# Document Processing
pypdf==3.17.0
docx2txt==0.8
python-pptx==0.6.21

# Utilities
requests==2.31.0
numpy==1.24.3
tqdm==4.66.1